Advection refactor and initial commit of B fields in fermi_circle
