+++ src/qimpy/transport/advection/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ["Vprime"]
+
+from ._advection import Vprime
